Refactoring, extract validation to a dedicated service, update deserializer to strictly only accept a string array.

+++ src/main/java/com/exercise/ordermanagement/dto/OrderResponse.java
@@ -3,10 +3,11 @@ package com.exercise.ordermanagement.dto;
 import com.exercise.ordermanagement.entity.Order;
 import com.exercise.ordermanagement.enums.OrderStatus;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data
 public class OrderResponse {
-    private Long id;
+    private Integer id;
     private Integer distance;
     private OrderStatus status;
